observe: workaround avoids multiple registeration
Reconnecting to Golioth before the registration process for observed
data completes will result in multiple observations to the same
endpoint.

Using the store settings for PSK/PSK-ID prompts a mandatory
reconnect that often triggers this issue. It's a known problem that will
be fixed in the SDK. This commit is a workaround until the changes are
released.

Signed-off-by: Mike Szczys <mike@golioth.io>